requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
requests
torch
torchvision
opencv-python
Pillow
tqdm
